+++ private/network/get_info.js
@@ -5,7 +5,7 @@ const rpc_call = require('./rpc_call')
 /**
  * Returns info regarding the EOSIO blockchain.
  * @param {object} arg Argument.
- * @param {string[]} arg.rpc_urls List of URL strings.
+ * @param {Array<string>} arg.rpc_urls List of URL strings.
  * @returns {object} Get info object.
  */
 async function get_info({ rpc_urls }) {