+++ public/index.mjs
@@ -1,3 +1,0 @@
-import smartql from './index.js'
-
-export default smartql