+++ test/index.test.mjs
@@ -1,4 +1,4 @@
-import { TestDirector } from 'test-director'
+import TestDirector from 'test-director'
 import wasm from './wasm.test.mjs'
 
 const tests = new TestDirector()